Constraints modify enabled/disabled status.
Check constraints recreate themselves when their expression differs.

HasColumns compares columns only by names.

Any HasProp is permissive by default if the LHS property is None.

When a CheckConstraint is disabled, it looses its reference to the index in the DB. Therefore only pull the index if there's anything there.

Parsed CheckConstraints needed to have the columns referenced in the expression set as the columns property.

A constrant is considered enabled if no status is specfied, rather than ignoring the property value.