A more general approach to dealing with dangling references in the cache.
In a previous commit, I tackled the need to refresh objects who referenced other
objects, but their own timestamp hasn't changed. This new approach eliminates
the need to explicitly find the latest timestamp, and instead invalidates and
refereshes all objects that reference an object that is actually changed in the
DB as indicated by it's change date. This extra invalidation only goes one level
deep in the dependency graph, so excessive refreshes are avoided. As a bonus,
some previous code was explititly handling the fact that we refresh all Columns
any time a Table changes, and now this process handles that too.