Warn when an INSERT statment underspecifies the columns required in a primary/unique key.
If the insert underspecifies the key, Precog will want to delete and reinsert the row every time because the specified columns are not a superset of any unique key, thus the _comp() tuple uses only the specified columns. This will not match the database's _comp() tuple as it will use whatever unique key is avaiable.

Now we warn about this case, and suggest what columns should be added to the insert statement to satisfy the unique key.